+++ src/Stack/Hoogle.hs
@@ -164,7 +164,7 @@ hoogleCmd (args,setup,rebuild,startServer) go = withBuildConfig go $ do
             Right hooglePath -> do
                 result <- withProcessContext menv
                         $ proc hooglePath ["--numeric-version"]
-                        $ tryAny . readProcessStdout_
+                        $ tryAny . fmap fst . readProcess_
                 let unexpectedResult got = Left $ T.concat
                         [ "'"
                         , T.pack hooglePath