+++ test/integration/tests/3787-internal-libs-with-no-main-lib/files/src-sublib/B.hs
@@ -1,0 +1,5 @@
+module B where
+
+-- | A function of the internal library
+funInternal :: Int -> Int
+funInternal = pred