+++ test/integration/tests/4105-test-coverage-of-internal-lib/files/src/Src.hs
@@ -1,0 +1,5 @@
+module Src where
+
+-- | A function of the main library
+funMainLib :: Int -> Int
+funMainLib = succ